requirements.txt
discord.py>=2.3.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
nodriver>=0.8.0
beautifulsoup4>=4.12.0
pandas>=2.0.0